Use bss section for sramptr
bss section is mapped to sram for both flash and qspi linker scripts, but
that's not the case for the sram lds.
We were using heap before, but then we saw that allocating through malloc
would overlap with our addresses.
In all cases, that's really a bss so a better section to describe it.