Refactor qspi init in board implementation
The comment was misleading: it's not that QSPI is already inited, it is
that it is not possible to run from QSPI and have it memory mapped at the
same time. To access QSPI, we need to have in the board, as while the
private implementation is global, it is set in a QSPI instance only in
Init.