Remove test/seed2-dfm-eval-euro from windows workflow
VCV needs -municode for build, which defines TCHAR as wchar_t while FatFs
also defines this type to char. So we have a global namespace name clash
we can't solve without changing FatFs implementation (which is available
as a dependency inside libdaisy). We therefore remove the test and
support for Windows.